upate 08.08.2019

+++ src/main/java/com/levermann/kennzahlen_fertig_review/Eigenkapitalrendite.java
@@ -73,7 +73,7 @@ final static Logger logger = Logger.getLogger(Eigenkapitalrendite.class);
 
                          if (lvsch.getUnternehmenname_Levermannschritte() == un.getUnternehmenname()  == true && retval > 0){
                              System.out.println("Richtig :D" + lvsch.getUnternehmenname_Levermannschritte() +" = " + un.getUnternehmenname() + "i = " + i);
-                             lvsch.setEigenkapitalrendite(1); }
+                             lvsch.setEigenkapitalrendite((float) 1); }
 
                      lvsch.setLevermannschrittAnalyseNameId(lvsch.getLevermannschrittAnalyseNameId());
                       // System.out.println("Unternehmen: " + un.getUnternehmenname() + " Levermannschritt: " + lvsch.getUnternehmenname_Levermannschritte()() );
@@ -94,7 +94,7 @@ final static Logger logger = Logger.getLogger(Eigenkapitalrendite.class);
 
                          if (lvsch1.getUnternehmenname_Levermannschritte() == un.getUnternehmenname()  == true && retval >= 0 ){
                              System.out.println("Richtig :D" + lvsch1.getUnternehmenname_Levermannschritte() +" = " + un.getUnternehmenname()  + "i = " + i);
-                             lvsch1.setEigenkapitalrendite(0); }
+                             lvsch1.setEigenkapitalrendite((float) 0); }
 
                      lvsch1.setLevermannschrittAnalyseNameId(lvsch1.getLevermannschrittAnalyseNameId());
 
@@ -111,7 +111,7 @@ final static Logger logger = Logger.getLogger(Eigenkapitalrendite.class);
 
                          if (lvsch1.getUnternehmenname_Levermannschritte() == un.getUnternehmenname()  == true) {
                              System.out.println("Richtig :D" + lvsch1.getUnternehmenname_Levermannschritte() + " = " + un.getUnternehmenname()  + "i = " + i);
-                             lvsch1.setEigenkapitalrendite(-1);}
+                             lvsch1.setEigenkapitalrendite((float) -1);}
 
                          lvsch1.setLevermannschrittAnalyseNameId(lvsch1.getLevermannschrittAnalyseNameId());
                          //  System.out.println("Fall 3 : yea LevermannschrittAnalyseNameId:  " + lvsch1.getLevermannschrittAnalyseNameId() + " AM: " + lvsch1.getGewinnrevision());
