+++ src/main/java/com/levermann/kennzahlen_fertig_review/KursheuteggKursvor1Jahr.java
@@ -73,7 +73,7 @@ public class KursheuteggKursvor1Jahr {
 
                         if (lvsch.getUnternehmenname_Levermannschritte() == un.getUnternehmenname()  == true && i >= 0.05 == true) {
                             System.out.println("Richtig :D" + lvsch.getUnternehmenname_Levermannschritte() + " = " + un.getUnternehmenname()  + "i = " + i);
-                            lvsch.setKursverlauf12Monate(1);
+                            lvsch.setKursverlauf12Monate((float) 1);
                         }
 
                         lvsch.setLevermannschrittAnalyseNameId(lvsch.getLevermannschrittAnalyseNameId());
@@ -92,7 +92,7 @@ public class KursheuteggKursvor1Jahr {
 
                         if (lvsch1.getUnternehmenname_Levermannschritte() == un.getUnternehmenname()  == true && i <= -0.05 == true) {
                             System.out.println("Richtig :D" + lvsch1.getUnternehmenname_Levermannschritte() + " = " + un.getUnternehmenname()  + "i = " + i);
-                            lvsch1.setKursverlauf12Monate(-1);
+                            lvsch1.setKursverlauf12Monate((float) -1);
                         }
 
                         lvsch1.setLevermannschrittAnalyseNameId(lvsch1.getLevermannschrittAnalyseNameId());
@@ -111,7 +111,7 @@ public class KursheuteggKursvor1Jahr {
 
                         if (lvsch1.getUnternehmenname_Levermannschritte() == un.getUnternehmenname()  == true && i > 0.005 == false && i < -0.05 == false) {
                             System.out.println("Richtig :D" + lvsch1.getUnternehmenname_Levermannschritte() + " = " + un.getUnternehmenname()  + "i = " + i);
-                            lvsch1.setKursverlauf12Monate(0);
+                            lvsch1.setKursverlauf12Monate((float) 0);
                         }
 
                         lvsch1.setLevermannschrittAnalyseNameId(lvsch1.getLevermannschrittAnalyseNameId());
